refactor: refactor wordcloud-shape to support renderSync
